#155 add CardInfo (organization code)
