+++ src/views/PersonalHistory/styles.js
@@ -8,7 +8,7 @@ const useStyles = makeStyles(theme => ({
   },
   header: {
     display: 'flex',
-    flexDirection: 'column'
+    flexDirection: 'column',
   },
   titleWrapper: {
     display: 'flex',
@@ -37,37 +37,11 @@ const useStyles = makeStyles(theme => ({
     minHeight: 800,
     marginTop: 40,
     paddingLeft: 40,
-    paddingTop: 24
+    paddingTop: 24,
   },
   formGroup: {
-    marginBottom: 24
+    marginBottom: 24,
   },
-  cardInfo: {
-    padding: 16,
-    width: 570,
-    marginBottom: 10,
-  },
-  title: {
-    fontWeight: 'bold',
-    color: theme.palette.primary.main,
-    letterSpacing: 1.25,
-    textTransform: 'uppercase',
-    fontSize: 16
-  },
-  label: {
-    fontWeight: 'bold',
-    fontSize: 14,
-    marginRight: 4
-  },
-  description: {
-    fontSize: 14
-  },
-  titleContainer: {
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center'
-  }
 }));
 
 export default useStyles;
